i should commit more frequently

+++ src/main/java/loqor/ait/core/blocks/ExteriorBlock.java
@@ -62,7 +62,7 @@ import java.util.function.ToIntFunction;
 @SuppressWarnings("deprecation")
 public class ExteriorBlock extends Block implements BlockEntityProvider, ICantBreak, Waterloggable {
 
-	public static final int MAX_ROTATION_INDEX = RotationPropertyHelper.getMax();
+	public static final byte MAX_ROTATION_INDEX = (byte) RotationPropertyHelper.getMax();
 	private static final int MAX_ROTATIONS = MAX_ROTATION_INDEX + 1;
 	public static final IntProperty ROTATION = Properties.ROTATION;
 	public static final IntProperty LEVEL_9 = Properties.LEVEL_15;
